+++ containers/Compute/constants/index.js
@@ -538,7 +538,7 @@ export const STORAGE_TYPES = {
     },
     io2: {
       label: i18n.t('common.storage.io', ['io2']),
-      value: 'io1',
+      value: 'io2',
       min: 4,
       max: 16384,
       sysMin: 4,
