+++ containers/Compute/sections/ServerNetwork/NetworkConfig.vue
@@ -35,7 +35,7 @@
             remote
             show-sync
             :item.sync="item.network"
-            :isDefaultSelect="i === 0"
+            :isDefaultSelect="canDefaultSelect && i === 0"
             :need-params="true"
             :params="{ ...networkParamsC, $t: item.key }"
             :mapper="networkResourceMapper"
@@ -51,6 +51,7 @@
       <div :class="{ 'd-flex ml-1' : isBigScreen && !isDialog }">
         <!-- 高级 -->
         <template v-if="showAdvanced">
+          <!-- ip -->
           <template v-if="isSupportIPv4(item) && !(isSupportIPv6(item) && item.ipv6Mode === 'only' && item.requireIpv6)">
             <template v-if="item.ipShow">
               <a-form-item class="mb-0" style="display:inline-block" :wrapperCol="{ span: 24 }">
@@ -62,6 +63,7 @@
               <a-button type="link" class="mr-1 mt-1" :disabled="ipsDisabled" @click="triggerShowIp(item)">{{$t('compute.text_198')}}</a-button>
             </a-tooltip>
           </template>
+          <!-- mac -->
           <template v-if="showMacConfig">
             <template v-if="item.macShow">
               <a-form-item class="mb-0" style="display:inline-block" :wrapperCol="{ span: 24 }">
@@ -77,6 +79,7 @@
               <a-button type="link" class="mr-1 mt-1" :disabled="ipsDisabled" @click="triggerShowMac(item)">{{$t('compute.mac_config')}}</a-button>
             </a-tooltip>
           </template>
+          <!-- 透传设备 -->
           <template v-if="showDeviceConfig">
             <template v-if="item.deviceShow">
               <a-form-item class="mb-0" style="display:inline-block" :wrapperCol="{ span: 24 }">
@@ -90,6 +93,7 @@
             </template>
             <a-button v-else type="link" class="mr-1 mt-1" @click="triggerShowDevice(item)">{{ $t('compute.config_transparent_net') }}</a-button>
           </template>
+          <!-- ipv6 -->
           <template>
             <a-form-item class="mb-0" style="display:inline-block" :wrapperCol="{ span: 24 }" v-if="isSupportIPv6(item) && isSupportIPv4(item)">
               <div class="d-flex align-items-center">
@@ -222,6 +226,7 @@ export default {
       networkOpts: [],
       screenWidth: document.body.clientWidth,
       showAdvanced: false,
+      canDefaultSelect: true,
     }
   },
   computed: {
@@ -285,6 +290,78 @@ export default {
     window.removeEventListener('resize', this.onResize)
   },
   methods: {
+    initData (data) {
+      this.canDefaultSelect = false
+      this.networkList = data.map(item => {
+        const obj = {
+          ...item,
+          key: uuid(),
+          network: { id: item.network },
+          vpc: { id: item.vpc },
+          ipShow: !!item.address,
+          ipv6Show: false,
+          requireIpv6: false,
+          ipv6Mode: 'all',
+          macShow: false,
+          deviceShow: false,
+          ip: item.address,
+        }
+        if (item.address) {
+          obj.ipShow = true
+          this.showAdvanced = true
+        }
+        if (item.mac) {
+          obj.macShow = true
+          this.showAdvanced = true
+        }
+        if (item.require_ipv6) {
+          obj.requireIpv6 = true
+          obj.ipv6Mode = 'all'
+          this.showAdvanced = true
+        }
+        if (item.strict_ipv6 && item.require_ipv6) {
+          obj.ipv6Mode = 'only'
+        }
+        if (item.address6) {
+          obj.ipv6Show = true
+          this.showAdvanced = true
+        }
+        if (item.sriov_device && item.sriov_device.model) {
+          obj.deviceShow = true
+          this.showAdvanced = true
+        }
+        return obj
+      })
+      this.$nextTick(() => {
+        this.form.fc.setFieldsValue({
+          [this.decorator.vpcs(this.networkList[0].key)[0]]: this.networkList[0].vpc.id,
+        })
+        for (const item of this.networkList) {
+          const value = {}
+          value[this.decorator.networks(item.key)[0]] = item.network.id
+          if (item.address) {
+            value[this.decorator.ips(item.key, item.network.id)[0]] = item.address
+          }
+          if (item.mac) {
+            value[this.decorator.macs(item.key, item.network.id)[0]] = item.mac
+          }
+          value[this.decorator.ipv6s(item.key, item.network.id)[0]] = item.requireIpv6
+          value[this.decorator.ipv6_mode(item.key, item.network.id)[0]] = item.ipv6Mode
+          if (item.address6) {
+            value[this.decorator.ips6(item.key, item.network.id)[0]] = item.address6.replace(this.getIpv6Prefix(item.address6), '')
+          }
+          if (item.sriov_device && item.sriov_device.model) {
+            value[this.decorator.devices(item.key)[0]] = item.sriov_device.model
+          }
+          setTimeout(() => {
+            this.form.fc.setFieldsValue(value)
+          }, 2000)
+          setTimeout(() => {
+            this.form.fc.setFieldsValue(value)
+          }, 4000)
+        }
+      })
+    },
     getVpcTag (data) {
       if (!data.cidr_block) return data.name
       return `${data.name}（${data.cidr_block}）`
@@ -337,6 +414,7 @@ export default {
         macShow: false,
         deviceShow: false,
         key: uid,
+        ip: '',
       }
       if (this.vpcObj) {
         data.vpc = this.vpcObj
@@ -448,12 +526,16 @@ export default {
       }
       return false
     },
-    fetchVpcSuccessHandle (data, item) {
-      item.vpc = data[0]
+    fetchVpcSuccessHandle (data = [], item) {
+      let target = data[0] || {}
+      if (item.vpc?.id && data.some(i => i.id === item.vpc?.id)) {
+        target = data.filter(i => i.id === item.vpc?.id)[0]
+      }
+      item.vpc = target
       this.$nextTick(() => {
-        this.form.fc.setFieldsValue({ [`vpcs[${item.key}]`]: data?.[0]?.key, vpcs: { [item.key]: data?.[0]?.key } })
+        this.form.fc.setFieldsValue({ [`vpcs[${item.key}]`]: target?.key, vpcs: { [item.key]: target?.key } })
       })
-      this.vpcSelectChange([data[0]], 0, item)
+      this.vpcSelectChange([target], 0, item)
       // this.fetchNetworkOpts(this.networkParamsC, item)
     },
     fetchNetworkSuccessHandle (data, item) {
