+++ containers/Compute/views/vminstance/dialogs/Restart.vue
@@ -7,8 +7,8 @@
           <div>{{$t('compute.text_1234')}}</div>
         </template>
       </a-alert>
-      <dialog-selected-tips :name="$t('dictionary.server')" :count="params.data.length" :action="action" />
-      <vxe-grid class="mb-2" :data="params.data" :columns="columns" />
+      <dialog-selected-tips :name="params.name || $t('dictionary.server')" :count="dataList.length" :action="action" />
+      <vxe-grid class="mb-2" :data="dataList" :columns="columns" />
       <a-form :form="form.fc" hideRequiredMark>
         <a-form-item :label="$t('compute.text_1041')" v-bind="formItemLayout" v-if="isOpenWorkflow">
           <a-input v-decorator="decorators.reason" :placeholder="$t('compute.text_1105')" />
@@ -31,14 +31,17 @@ import DialogMixin from '@/mixins/dialog'
 import WindowsMixin from '@/mixins/windows'
 import WorkflowMixin from '@/mixins/workflow'
 import { BATCH_OPERATE_SERVERS_MAX } from '@/constants/workflow'
+import { getNameDescriptionTableColumn, getStatusTableColumn } from '@/utils/common/tableColumn'
 
 export default {
   name: 'VmRestartDialog',
   mixins: [DialogMixin, WindowsMixin, WorkflowMixin],
   data () {
+    const { type, formData = {} } = this.params
     return {
       loading: false,
-      action: this.$t('compute.text_274'),
+      action: type === 'modifyWorkflow' ? this.$t('common.modify_workflow') + `(${this.$t('compute.text_274')})` : this.$t('compute.text_274'),
+      type,
       form: {
         fc: this.$form.createForm(this),
       },
@@ -46,13 +49,13 @@ export default {
         reason: [
           'reason',
           {
-            initialValue: '',
+            initialValue: formData.reason || '',
           },
         ],
         autoStart: [
           'autoStart',
           {
-            initialValue: false,
+            initialValue: formData.is_force || false,
             valuePropName: 'checked',
           },
         ],
@@ -65,11 +68,15 @@ export default {
           span: 3,
         },
       },
+      dataList: [],
     }
   },
   computed: {
     ...mapGetters(['userInfo']),
     columns () {
+      if (this.type === 'modifyWorkflow') {
+        return [getNameDescriptionTableColumn(), getStatusTableColumn({ statusModule: 'server' })]
+      }
       const showFields = ['name', 'ip', 'instance_type']
       return this.params.columns.filter((item) => { return showFields.includes(item.field) })
     },
@@ -77,9 +84,22 @@ export default {
       return this.checkWorkflowEnabled(this.WORKFLOW_TYPES.APPLY_SERVER_RESTART)
     },
   },
+  created () {
+    this.initData()
+  },
   methods: {
+    async initData () {
+      if (this.type === 'modifyWorkflow' && this.params.ids) {
+        const list = await new this.$Manager('servers', 'v2').batchGet({
+          id: this.params.ids,
+        })
+        this.dataList = list.data?.data || []
+      } else {
+        this.dataList = this.params.data || []
+      }
+    },
     async doRestartSubmit (values) {
-      const ids = this.params.data.map(item => item.id)
+      const ids = this.dataList.map(item => item.id)
       return this.params.onManager('batchPerformAction', {
         id: ids,
         steadyStatus: 'ready',
@@ -96,12 +116,12 @@ export default {
       try {
         const values = await this.form.fc.validateFields()
         if (this.isOpenWorkflow) {
-          if (this.params.data.length > BATCH_OPERATE_SERVERS_MAX) {
+          if (this.dataList.length > BATCH_OPERATE_SERVERS_MAX) {
             this.$message.error(this.$t('compute.workflow.operate_servers_check.message', [this.$t('compute.text_274'), BATCH_OPERATE_SERVERS_MAX]))
             this.loading = false
             return
           }
-          const projects = new Set(this.params.data.map(item => item.tenant_id))
+          const projects = new Set(this.dataList.map(item => item.tenant_id))
           if (projects.size > 1) {
             this.$message.error(this.$t('compute.text_1348'))
             this.loading = false
@@ -118,20 +138,25 @@ export default {
       }
     },
     async handleRestartByWorkflowSubmit (values) {
-      const ids = this.params.data.map(item => item.id)
+      const ids = this.dataList.map(item => item.id)
       const params = {
         is_force: values.autoStart,
       }
       const variables = {
-        project: this.params.data[0].tenant_id,
-        project_domain: this.params.data[0].domain_id,
+        project: this.dataList[0].tenant_id,
+        project_domain: this.dataList[0].domain_id,
         process_definition_key: this.WORKFLOW_TYPES.APPLY_SERVER_RESTART,
         initiator: this.userInfo.id,
         ids: ids.join(','),
         description: values.reason,
         paramter: JSON.stringify(params),
       }
-      await this.createWorkflow(variables)
+      if (this.type === 'modifyWorkflow') {
+        await this.updateWorkflow(variables, this.params.workflow)
+        this.params.success && this.params.success()
+      } else {
+        await this.createWorkflow(variables)
+      }
       this.$message.success(this.$t('compute.text_1214'))
       this.$router.push('/workflow')
     },
