+++ containers/Compute/views/vminstance/dialogs/Start.vue
@@ -2,8 +2,8 @@
   <base-dialog @cancel="cancelDialog">
     <div slot="header">{{action}}</div>
     <div slot="body">
-      <dialog-selected-tips :name="$t('dictionary.server')" :count="params.data.length" :action="action" />
-      <dialog-table :data="params.data" :columns="columns" />
+      <dialog-selected-tips :name="params.name || $t('dictionary.server')" :count="dataList.length" :action="action" />
+      <dialog-table :data="dataList" :columns="columns" />
       <a-form
         :form="form.fc">
         <a-form-item :label="$t('compute.text_1041')" v-bind="formItemLayout" v-if="isOpenWorkflow">
@@ -11,7 +11,7 @@
         </a-form-item>
         <a-form-item class="mb-0" v-show="canAutoPrepaid">
           <a-checkbox
-          :checked="form.fd.auto_prepaid"
+          v-model="form.fd.auto_prepaid"
           @change="autoPrepaidChange">
             {{$t('compute.change_to_prepaid')}}
           </a-checkbox>
@@ -31,32 +31,35 @@ import DialogMixin from '@/mixins/dialog'
 import WindowsMixin from '@/mixins/windows'
 import WorkflowMixin from '@/mixins/workflow'
 import { BATCH_OPERATE_SERVERS_MAX } from '@/constants/workflow'
+import { getNameDescriptionTableColumn, getStatusTableColumn } from '@/utils/common/tableColumn'
 
 export default {
   name: 'VmStartDialog',
   mixins: [DialogMixin, WindowsMixin, WorkflowMixin],
   data () {
+    const { type, formData = {} } = this.params
     return {
       loading: false,
-      action: this.$t('compute.text_272'),
+      type,
+      action: type === 'modifyWorkflow' ? this.$t('common.modify_workflow') + `(${this.$t('compute.text_272')})` : this.$t('compute.text_272'),
       form: {
         fc: this.$form.createForm(this),
         fd: {
-          auto_prepaid: false,
+          auto_prepaid: formData.auto_prepaid || false,
         },
       },
       decorators: {
         reason: [
           'reason',
           {
-            initialValue: '',
+            initialValue: formData.reason || '',
           },
         ],
         auto_prepaid: [
           'stopPaying',
           {
             valuePropName: 'checked',
-            initialValue: false,
+            initialValue: formData.auto_prepaid || false,
           },
         ],
       },
@@ -68,20 +71,24 @@ export default {
           span: 3,
         },
       },
+      dataList: [],
     }
   },
   computed: {
     ...mapGetters(['userInfo']),
     columns () {
+      if (this.type === 'modifyWorkflow') {
+        return [getNameDescriptionTableColumn(), getStatusTableColumn({ statusModule: 'server' })]
+      }
       const showFields = ['name', 'ip', 'instance_type']
       return this.params.columns.filter((item) => { return showFields.includes(item.field) })
     },
     // 腾讯云、阿里云、火山云的按量付费机器，关机可停止付费
     // 腾讯云、阿里云包年包月机器，关机停止付费会转为按量付费机器
     canAutoPrepaid () {
-      return this.params.data.every(item => {
+      return this.dataList.every(item => {
         return ['qcloud', 'aliyun'].includes(item.brand.toLocaleLowerCase())
-      }) && this.params.data.some(item => {
+      }) && this.dataList.some(item => {
         return item.billing_type === 'postpaid'
       })
     },
@@ -89,13 +96,26 @@ export default {
       return this.checkWorkflowEnabled(this.WORKFLOW_TYPES.APPLY_SERVER_START)
     },
   },
+  created () {
+    this.initData()
+  },
   methods: {
+    async initData () {
+      if (this.type === 'modifyWorkflow' && this.params.ids) {
+        const list = await new this.$Manager('servers', 'v2').batchGet({
+          id: this.params.ids,
+        })
+        this.dataList = list.data?.data || []
+      } else {
+        this.dataList = this.params.data || []
+      }
+    },
     async doShutDownSubmit () {
       const data = {}
       if (this.form.fd.auto_prepaid) {
         data.auto_prepaid = true
       }
-      const ids = this.params.data.map(item => item.id)
+      const ids = this.dataList.map(item => item.id)
       return this.params.onManager('batchPerformAction', {
         id: ids,
         steadyStatus: 'running',
@@ -109,12 +129,12 @@ export default {
       this.loading = true
       try {
         if (this.isOpenWorkflow) {
-          if (this.params.data.length > BATCH_OPERATE_SERVERS_MAX) {
+          if (this.dataList.length > BATCH_OPERATE_SERVERS_MAX) {
             this.$message.error(this.$t('compute.workflow.operate_servers_check.message', [this.$t('compute.text_273'), BATCH_OPERATE_SERVERS_MAX]))
             this.loading = false
             return
           }
-          const projects = new Set(this.params.data.map(item => item.tenant_id))
+          const projects = new Set(this.dataList.map(item => item.tenant_id))
           if (projects.size > 1) {
             this.$message.error(this.$t('compute.text_1348'))
             this.loading = false
@@ -136,18 +156,24 @@ export default {
       this.form.fd.auto_prepaid = checked
     },
     async handleShutDownByWorkflowSubmit () {
-      const ids = this.params.data.map(item => item.id)
+      const ids = this.dataList.map(item => item.id)
       const values = await this.form.fc.validateFields()
       const params = {}
       if (this.form.fd.auto_prepaid) params.auto_prepaid = true
       const variables = {
-        project: this.params.data[0].tenant_id,
-        project_domain: this.params.data[0].domain_id,
+        project: this.dataList[0].tenant_id,
+        project_domain: this.dataList[0].domain_id,
         process_definition_key: this.WORKFLOW_TYPES.APPLY_SERVER_START,
         initiator: this.userInfo.id,
         ids: ids.join(','),
         description: values.reason,
         paramter: JSON.stringify(params),
+      }
+      if (this.type === 'modifyWorkflow') {
+        await this.updateWorkflow(variables, this.params.workflow)
+        this.params.success && this.params.success()
+      } else {
+        await this.createWorkflow(variables)
       }
       await this.createWorkflow(variables)
       this.$message.success(this.$t('compute.text_1214'))
