+++ src/locales/ja-JP.json
@@ -2910,7 +2910,8 @@
     "menu_iam_security": "認証とセキュリティ",
     "menu_system_config": "システム設定",
     "global_topology": "グローバルトポロジ",
-    "k8s": "K8S"
+    "k8s": "K8S",
+    "server_container": "コンテナホスト"
   },
   "originDictionary": {},
   "res": {
@@ -6227,5 +6228,13 @@
   "common.pem_certificate.placeholder": "証明書を以下のように貼り付けてください：\n-----BEGIN CERTIFICATE-----\nQmFzZTY0IGVuY29kZWQgY2VydGlmaWNhdGUgZGF0YSBleGlzdHMgaGVyZQ==\n-----END CERTIFICATE-----",
   "common.pem_private_key.placeholder": "秘密鍵を以下のように貼り付けてください：\n-----BEGIN PRIVATE KEY-----\nQmFzZTY0IGVuY29kZWQgc2VjcmV0IGtleSBleGlzdHMgaGVyZQ==\n-----END PRIVATE KEY-----",
   "common.switch_to_input": "入力に切り替え",
-  "common.switch_to_select": "選択に切り替え"
+  "common.switch_to_select": "選択に切り替え",
+  "common.modify_workflow": "製造オーダの変更",
+  "common.current_quota": "現在のクォータ",
+  "common.new_quota": "申請割当",
+  "common.workflow_disabled_tip": "このタイプの製造オーダはこのオペレーションをサポートしていません",
+  "common.workflow_disabled_tip2": "承認されていない製造オーダはこの操作をサポートしています",
+  "common.workflow_disabled_tip3": "このタイプの製造オーダは承認プロセスをオープンしていません。製造オーダ構成でオープンしてください",
+  "common.add_resource_from_shopcart": "カートからリソースを追加",
+  "common.delete_resource_from_shopcart_tip": "この操作は製造オーダからリソースを削除します。この操作を確認しますか?"
 }
